panw-takehome/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "panw-takehome"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8",
    "requests>=2.31",
    "allure-pytest>=2.13",
    "pyyaml>=6",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
